requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
altair=4.2.0=pypi_0
appdirs=1.4.4=pyhd3eb1b0_0
asttokens=2.0.5=pyhd3eb1b0_0
async_generator=1.10=pyhd3eb1b0_0
attrs=22.1.0=py39haa95532_0
automat=22.10.0=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
backports-zoneinfo=0.2.1=pypi_0
beautifulsoup4=4.11.1=py39haa95532_0
blas=1.0=mkl
blinker=1.5=pypi_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
brotlipy=0.7.0=py39h2bbff1b_1003
bs4=4.11.1=hd8ed1ab_0
ca-certificates=2023.01.10=haa95532_0
cachetools=5.2.0=pypi_0
certifi=2022.12.7=py39haa95532_0
cffi=1.15.1=py39h2bbff1b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py39haa95532_0
cloudpickle=2.0.0=pyhd3eb1b0_0
colorama=0.4.6=py39haa95532_0
commonmark=0.9.1=pypi_0
constantly=15.1.0=pypi_0
cryptography=38.0.4=py39h21b164f_0
cssselect=1.2.0=pypi_0
cycler=0.11.0=pyhd3eb1b0_0
decorator=5.1.1=pyhd3eb1b0_0
entrypoints=0.4=pypi_0
executing=0.8.3=pyhd3eb1b0_0
fftw=3.3.9=h2bbff1b_1
filelock=3.9.0=pypi_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=ha860e81_0
gitdb=4.0.10=pypi_0
gitpython=3.1.29=pypi_0
h11=0.12.0=pyhd3eb1b0_0
hyperlink=21.0.0=pypi_0
icc_rt=2022.1.0=h6049295_2
idna=3.4=py39haa95532_0
importlib-metadata=5.2.0=pypi_0
importlib-resources=5.10.1=pypi_0
incremental=22.10.0=pypi_0
intel-openmp=2021.4.0=haa95532_3556
ipykernel=5.3.4=py39h7b7c402_0
ipython=8.9.0=py39haa95532_0
itemadapter=0.7.0=pypi_0
itemloaders=1.0.6=pypi_0
jedi=0.18.1=py39haa95532_1
jinja2=3.1.2=pypi_0
jmespath=1.0.1=pypi_0
joblib=1.1.1=py39haa95532_0
jpeg=9e=h2bbff1b_0
jsonschema=4.17.3=pypi_0
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_core=4.11.2=py39haa95532_0
kiwisolver=1.4.4=py39hd77b12b_0
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libdeflate=1.8=h2bbff1b_5
libffi=3.4.2=hd77b12b_6
libiconv=1.16=h2bbff1b_2
libpng=1.6.37=h2a8f88b_0
libsodium=1.0.18=h62dcd97_0
libtiff=4.5.0=h6c2663c_1
libwebp=1.2.4=h2bbff1b_0
libwebp-base=1.2.4=h2bbff1b_0
libxml2=2.9.14=h0ad7f3c_0
libxslt=1.1.35=h2bbff1b_0
lxml=4.9.1=py39h1985fb9_0
lz4-c=1.9.4=h2bbff1b_0
markupsafe=2.1.1=pypi_0
matplotlib-base=3.5.3=py39hd77b12b_0
matplotlib-inline=0.1.6=py39haa95532_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py39h2bbff1b_0
mkl_fft=1.3.1=py39h277e83a_0
mkl_random=1.2.2=py39hf11a4ad_0
munkres=1.1.4=py_0
mysql-connector-python=8.0.31=pypi_0
numpy=1.24.0=pypi_0
numpy-base=1.23.5=py39h4da318b_0
openssl=1.1.1t=h2bbff1b_0
outcome=1.1.0=pyhd3eb1b0_0
packaging=22.0=py39haa95532_0
pandas=1.2.4=py39hd77b12b_0
parsel=1.7.0=pypi_0
parso=0.8.3=pyhd3eb1b0_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.3.0=pypi_0
pip=22.3.1=py39haa95532_0
pkgutil-resolve-name=1.3.10=pypi_0
pooch=1.4.0=pyhd3eb1b0_0
prompt-toolkit=3.0.36=py39haa95532_0
protego=0.2.1=pypi_0
protobuf=3.20.1=pypi_0
pure_eval=0.2.2=pyhd3eb1b0_0
pyarrow=11.0.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydeck=0.8.0=pypi_0
pydispatcher=2.0.6=pypi_0
pygments=2.13.0=pypi_0
pympler=1.0.1=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py39haa95532_0
pyrsistent=0.19.2=pypi_0
pysocks=1.7.1=py39haa95532_0
python=3.9.15=h6244533_2
python-dateutil=2.8.2=pyhd3eb1b0_0
python-dotenv=0.21.0=py39haa95532_0
pytz=2022.7=py39haa95532_0
pytz-deprecation-shim=0.1.0.post0=pypi_0
pywin32=305=py39h2bbff1b_0
pyzmq=23.2.0=py39hd77b12b_0
queuelib=1.6.2=pypi_0
requests=2.28.1=py39haa95532_0
requests-file=1.5.1=pypi_0
rich=12.6.0=pypi_0
scapy=2.5.0=pyhd8ed1ab_0
scikit-learn=1.2.0=py39hd77b12b_0
scipy=1.10.0=py39h321e85e_0
scrapy=2.7.1=pypi_0
selenium=4.7.2=pyhd8ed1ab_0
semver=2.13.0=pypi_0
service-identity=21.1.0=pypi_0
setuptools=65.6.3=py39haa95532_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
sniffio=1.2.0=py39haa95532_1
sortedcontainers=2.4.0=pyhd3eb1b0_0
soupsieve=2.3.2.post1=py39haa95532_0
spyder-kernels=2.2.0=py39haa95532_0
sqlite=3.40.1=h2bbff1b_0
stack_data=0.2.0=pyhd3eb1b0_0
streamlit=1.18.1=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h2bbff1b_0
tldextract=3.4.0=pypi_0
toml=0.10.2=pypi_0
toolz=0.12.0=pypi_0
tornado=6.2=py39h2bbff1b_0
tqdm=4.64.1=py39haa95532_0
traitlets=5.7.1=py39haa95532_0
trio=0.19.0=py39haa95532_0
trio-websocket=0.9.2=pyhd8ed1ab_0
twisted=22.10.0=pypi_0
twisted-iocpsupport=1.0.2=pypi_0
typing-extensions=4.5.0=pypi_0
tzdata=2022.7=pypi_0
tzlocal=4.2=pypi_0
urllib3=1.26.14=py39haa95532_0
validators=0.20.0=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
w3lib=2.1.1=pypi_0
watchdog=2.2.0=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webdriver-manager=3.8.5=pyhd8ed1ab_0
wheel=0.37.1=pyhd3eb1b0_0
win_inet_pton=1.1.0=py39haa95532_0
wincertstore=0.2=py39haa95532_2
wsproto=1.2.0=pyhd8ed1ab_0
xz=5.2.10=h8cc25b3_1
zeromq=4.3.4=hd77b12b_0
zipp=3.11.0=pypi_0
zlib=1.2.13=h8cc25b3_0
zope-interface=5.5.2=pypi_0
zstd=1.5.2=h19a0ad4_0
